% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_weight_formula.R
\name{weight_formula}
\alias{weight_formula}
\title{weight_formula}
\usage{
weight_formula(
  weight,
  height,
  sex,
  weight_unit = c("kg", "lbs"),
  formula = c("IBW", "AJBW", "LBW")
)
}
\arguments{
\item{weight}{correspond to the total body weigt in kg or lbs}

\item{height}{correspond to the patient height (only in cm)}

\item{sex}{correspond to the patient sex}

\item{weight_unit}{should be either kg or lbs}

\item{formula}{correspond to the formula used. Can take value in c(IBW, AJBW, LBW)}
}
\description{
this function calculated different weight indicator, such as
the Ideal Body weigh (IBW) or the adjusted bodyweight (AJBW)
}
