% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pta_plot.R
\name{plot.cfr}
\alias{plot.cfr}
\title{plot.cfr}
\usage{
\method{plot}{cfr}(data)
}
\arguments{
\item{data}{take a data.frame with the following columns: dose, cfr, toxicity_proportion}
}
\description{
Plot the cumulative fraction rate based on eucast value if available
}
\author{
Romain Garreau
}
